requirements.txt
langgraph
langchain
langchain-openai
langchain-community
redis
chromadb
tiktoken
python-dotenv
pydantic
